Support horizontal panning with middle button or Ctrl+left button
Allow panning with the mouse with middle button drag or Ctrl+left button
drag. Change the cursor to 'grabbing' while panning is ongoing.

Change the cursor to 'grabbing' when only Ctrl is pressed to indicate
that left button can be used to pan.

Restore the cursor to 'default' when panning is ended by releasing the
mouse button that initiated it, or when the conditions to initiate
Ctrl+left button panning are no longer met.

Prevent time selection with any other button than the left button or
when the Ctrl or Alt key is pressed.

Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>