Use crosshair cursor for time selection
Change the cursor to 'crosshair' while time selection is ongoing.

Change the cursor to 'crosshair' when only Shift is pressed to indicate
that left button can be used to extend time selection.

Restore the cursor to 'default' when time selection is ended by
releasing the left mouse button, or when the conditions to initiate
Shift+left button time selection are no longer met.

Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>