Fix mouse zooming when mouse released outside time graph
The x-coordinate of the mouseUp event is in the coordinates of the
element under the cursor at the time, which is not necessarily the time
graph coordinates.

Since the mouseEndX variable is updated (correctly) while the mouse is
being moved, just rely on that, and do not update it at mouseUp.

Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>