Center x-axis labels
Currently the labels of the x-axis are aligned to the left, 5 pixels
after their corresponding tick in the axis.

This commit centers the labels according to their ticks. The position of
the text is calculated by using the position of the tick, minus half the
width of the text.

Fixes [issue 440 in theia-ide/theia-trace-extension](https://github.com/theia-ide/theia-trace-extension/issues/440).

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>