Use BigInt for timestamps
Timestamps and time durations are changed to be of type 'bigint' instead
of 'number'.

This is done to avoid loss of precision in JavaScript 'number' type for
values above Number.MAX_SAFE_INTEGER (2^53 - 1).

The class BIMath is added to provide some BigInt math utility functions.

The method TimeGraphLayer.getPixels() is renamed to getPixel() since it
returns only one value.

In the time graph axis, fractional scale steps 1.5 and 2.5 are removed
to avoid non-equidistant ticks when zoomed in to nanosecond level, due
to BigInt integer rounding.

The minimum view range is set to 2 ns, to be able to zoom out of it and
to always have at least one time graph axis tick fully visible.

Make the minimum pan offset 1 ns when using keyboard 'A' or 'D' or
Shift-mouse wheel, for when the requested pixel offset represents less
than a nanosecond.

Refactor Ctrl+mouse drag horizontal panning to use the original start
position since the relative offset at each mouse move event can be less
than a nanosecond, preventing the panning.

Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>