Removed cached height and position in time graph state component
The height and position of a state is not fixed, it can be updated when
the view range is changed.

The time graph state component stores these both as field variables and
as part of the state options object. Only the variables in the options
are updated, the field variables remain to what was set at construction,
so the getters return obsolete values.

Remove the field variables and update the getters to return the latest
values from the options object.

Add a getter for the width.

Change-Id: I1987c2c2630d46eb27e64ca879e97fd942090824
Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>