Remove unnecessary gap states before blank states
There should be no gap state added before a blank state if its start
time is the same as the end time of the previous state.

There should be no gap state added before a blank state if the previous
state was not a blank state when both states connect on the same pixel.

Change-Id: I9d78f7b7429fc550500fa9dfec9a134b17e58691
Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>