Make gridlines aligned to rounded numbers
Add the offset to the time graph unit controller.

Use the offset to make sure that gridlines are set on rounded absolute
times, not rounded relative times to the start of the absolute range.

Calculate the maximum label width using the view range end time instead
of the view range length.

Fixes issue #54

Change-Id: I5ff00c57ae2859703c8eeb8b4db88bc1655d985f
Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>