Fix missing row updates
When the update flag in maybeFetchNewData() is set to true, it means all
rows must be refreshed (e.g. annotations might have changed). It was
only refreshing the visible rows, and when the other rows later became
visible they were considered valid and not refreshed.

Mark all existing rows as needing to be refreshed by setting their
providedModel object to undefined.

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>