Support coarse and fine update of time graph data
To enable this feature, provide a coarse resolution factor that is
higher than 1 to the TimeGraphChart constructor.

When a coarse resolution factor is provided, the first update of time
graph data is made using the coarse resolution. Once this update is
completed, a debounced fine resolution second update is triggered.

The resolution check against provided model is changed to require better
resolution, to prevent a coarse model from being fetched and used to
overwrite a previously fetched fine model.

When no row needs update at coarse resolution, try to update with fine
resolution. This allows the update to trigger when the time graph canvas
width is increased.

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>