Add support for configurable location of time axis ticks
This commit adds support for configurable location of time
axis ticks. Currently, the ticks can only be displayed above
the time labels in the time axis, which do not make sense UI-wise
when we have graphs that are displayed under the time axis.
This commit allows the ticks of the time axis to be displayed
dynamically either on top or under the time labels by adding
a configuration parameter to the TimeGraphAxisScale component,
so that developers can position the ticks correctly.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>