Improve rendering for text labels
Currently, when the timeline chart renders text labels for states, it deletes all
existing labels, then creates new labels, and adds them to the graph. This is not
efficient and makes rendering of the labels sluggish when zooming in and out of the
graph. This commit removes the code that delete the text labels objects on every
re-rerender, and updates the text content of the labels instead of re-creating them.

Fixes #225.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>