Cache label BitMapText objects in a map
This commit adds a map to the time-graph-state component as a cache
for the state labels. The cache is implemented as a map with the
text scaling value (display width / full label width) as the key and
the PIXI.BitMapText object as the value. With this change, the
timeline-chart can reuse display objects if they were not
changed, or if they were created recently.

The size of the cache is fixed, determined by the LABEL_CACHE_SIZE
constant. When updating the cache, the timeline-chart will remove the
oldest entry in the map. The reasoning behind this is that users are
more likely to zoom in/out contiguous zoom levels.

This commit improves the loading time when zooming in and out of the
timeline-chart.

Fixes #228.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>