Add ellipse and star to the character set of label text
This commit adds the ellipse (...) and star unicode characters
to the character set of the label text in the time-graph-font
controller. This would help display the ellipse symbol in the label
text (when the text is truncated) as a single character and takes
up the same space as a single character.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>