Make scaling state a handler of onScaleFactorChange event
Currently, scaling state happens when the view range changes. This is not
the expected behaviour. States should only be scaled when the scale factor
changes. This commit moves the function that scales the states from
the view range change handlers to the scale factor change handlers. This
will also help make better code isolation for unit tests.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>