Fix time graph glitches when zooming with active filter
- Wait until a new model is applied before setting the new world range
and resetting the scale to 1. This could lead to the world range being
changed without updating the world model coordinates.

- Move the world range fields and methods from the common unit
controller to the state controller, since each time graph layer can have
its own independent world range.

- Change method updateWorldRangeFromViewRange() to
computeWorldRangeFromViewRange() that calculates a new world range but
does not update the world range. This allows to calculate the world
range for a fetch row request without setting the world range yet.

- Do not reset the scale to 1 when a received model is discarded or
rejected. This could revert a view range change done by a user
operation.

- Do not update world range and scale during full search one-at-a-time
row updates. This avoids setting an obsolete world range when a row is
returned and accepted after the view range has been changed by a user
operation but before this has triggered a new request.

- Add a new field worldZoomFactor to the state controller. Update the
world zoom factor when a new world range is set, but do not update it
when the view range is zoomed.

- Use the world zoom factor to calculate world coordinates in
getWorldPixel(), instead of the view range zoom factor, but only if the
layer is scalable. For non-scalable layers, keep using the view range
zoom factor.

- Force update of rows that become visible and do not have their
provided model range equal to the current world range. These rows can
have obsolete world coordinates.

- Add the scale factor to the constructor and as a field of a time graph
state. Use this scale factor when rendering labels, and remove the scale
factor as a parameter of renderLabel() method. This makes sure the scale
factor is not reset when re-rendering a label.

- When scaling state labels, re-render the label even if the new scale
factor is 1. This makes sure labels are not left at an obsolete scale
factor after zooming in and out back to the original zoom level.

- Add await for calls to fetchRows otherwise the possible
Promise.reject() is logged as an uncaught error.

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>