[CI] Add release workflow/job and publish-latest job
The new release workflow uses action "pipe-cd/actions-gh-release". This
action uses file RELEASE in the root. When that file is modified in a PR,
and contains a new release tag, it will trigger a release upon the PR
being merged. The action will create a git tag in the repo and the
corresponding GitHub release, containing auto-generated release notes,
that can be later edited by a committer.

The new release will in turn trigger the new "publish-latest" job
(ci-cd workflow), which will publish the new version of the package to
npm.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>