Fix time graph filtering and collapsing
Only clear excluded rows if the EXCLUDED filtering expressions have
changed. This prevents unnecessarily redoing filtering of rows when only
the search string has changed.

Determine the list of rows that have only changed y-position, for
example due to collapsing of preceding rows. If these rows do not need
to be fetched because their current model is valid, then rebuild the
rows using their current model to update each child element's position.

If no rows at all need to be fetched but there were some moved rows,
perform a fetch to update the arrows which may have changed their start
and/or end position.

Remove unused code for updating and removing of gap states, as the
states and gaps are always added from scratch when updating a row.
Remove the reference to gap state components in the row model states.

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>