Update pixi.js-legacy to latest major version
pixi.js-legacy@5.x -> pixi.js-legacy@7.x

The reason we use the legacy pixi.js is because there
are environments where timeline-chart is used, where no
GPU acceleration is possible, so there needs to be a
fall-back, when pixi's  WebGL reenderer can't be used.
This fallback is the CanvasRenderer, only available in
legacy pixi at the moment.

Generally we need to be careful to import pixi-related
objects from `pixi.js-legacy` rather than the modular
'@pixi/*'.

Misc:
- fix example application that no longer started
- try to do a better clean-up after each test
- ran "yarn upgrade": currently zero vulnerability
reported by "yarn audit"!
- removed node 18 from ci-cd workflow, since one
updated dependency is no longer compatible with this EoL
version of node:
error undici@7.10.0: The engine "node" is incompatible with this module. Expected version ">=20.18.1". Got "18.20.8"

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>